@@ -12,7 +12,6 @@ exclude = [
 
 [tool.hatch.build.force-include]
 "bowtie/schemas/io-schema.json" = "bowtie/schemas/io-schema.json"
-"bowtie/templates/" = "bowtie/templates/"
 
 [project]
 name = "bowtie-json-schema"
@@ -50,8 +49,8 @@ dependencies = [
   "attrs>=22.2.0",
   "click",
   "github3.py",
-  "jinja2",
-  "jsonschema",
+  "jsonschema>=4.19.0",
+  "referencing>=0.30.0",
   "rich",
   "structlog",
   "trogon>=0.4.0",
@@ -71,13 +70,6 @@ Source = "https://github.com/bowtie-json-schema/bowtie"
 branch = true
 source = ["bowtie"]
 dynamic_context = "test_function"
-
-[tool.djlint]
-# H008: Single and double quotes don't mean the same thing in Jinja, so it's
-#       not always the case that the latter should be used unfortunately.
-# J018: url_for is a flask thing, not a Jinja2 native one.
-ignore = "H008,J018"
-indent = 2
 
 [tool.doc8]
 ignore = [
@@ -102,8 +94,8 @@ exclude = [
 
 [tool.ruff]
 line-length = 79
-target-version = "py38"
-select = ["ANN", "B", "D", "D204", "E", "F", "Q", "SIM", "UP", "W"]
+target-version = "py310"
+select = ["ANN", "B", "D", "D204", "E", "F", "Q", "RUF", "SIM", "UP", "W"]
 ignore = [
   # Wat, type annotations for self and cls, why is this a thing?
   "ANN101",
@@ -116,6 +108,8 @@ ignore = [
   "B008",
   # raise SomeException(...) is fine.
   "B904",
+  # There's no need for explicit strict, this is simply zip's default behavior.
+  "B905",
   # It's fine to not have docstrings for magic methods.
   "D105",
   # __init__ especially doesn't need a docstring
@@ -136,9 +130,6 @@ ignore = [
   "D407",
   # Plz spaces after section headers
   "D412",
-  # We support 3.8 and 3.9
-  "UP006",
-  "UP007",
 ]
 extend-exclude = ["json"]
 
@@ -147,7 +138,7 @@ docstring-quotes = "double"
 
 [tool.ruff.per-file-ignores]
 "docs/*" = ["ANN", "D"]
-"noxfile.py" = ["ANN", "D"]
-"tests/*" = ["ANN", "D"]
+"noxfile.py" = ["ANN", "D100"]
+"tests/*" = ["ANN", "D", "RUF012"]
 "implementations/*" = ["ANN", "D"]
 "bowtie/schemas/__init__.py" = ["D"]
