@@ -3,17 +3,8 @@ Hand crafted classes which should undoubtedly be autogenerated from the schema.
 """
 from __future__ import annotations
 
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    ClassVar,
-    Iterable,
-    Protocol,
-    Type,
-    TypeVar,
-    cast,
-)
+from collections.abc import Callable, Iterable
+from typing import TYPE_CHECKING, Any, ClassVar, Protocol, TypeVar, cast
 
 try:
     from typing import dataclass_transform
@@ -113,8 +104,9 @@ def command(
     name: str,
     Response: Callable[..., _R | None],
 ) -> Callable[[type], type[Command[_R]]]:
-    request_schema = {"$ref": f"#/$defs/command/$defs/{name}"}
-    response_schema = {"$ref": f"#/$defs/command/$defs/{name}/$defs/response"}
+    uri = f"https://bowtie.report/io-schema/{name}/"
+    request_schema = {"$ref": uri}
+    response_schema = {"$ref": f"{uri}response/"}
 
     def _command(cls: type) -> type[Command[_R]]:
         def to_request(
@@ -137,7 +129,7 @@ def command(
             validate(instance=instance, schema=response_schema)
             return Response(**instance)
 
-        cls = cast(Type[Command[_R]], cls)
+        cls = cast(type[Command[_R]], cls)
         cls.to_request = to_request
         cls.from_response = from_response
         return frozen(cls)
@@ -260,11 +252,8 @@ class CaseResult:
 
     @classmethod
     def from_dict(cls, data: Any, **kwargs: Any) -> CaseResult:
-        return cls(
-            results=[TestResult.from_dict(t) for t in data.pop("results")],
-            **data,
-            **kwargs,
-        )
+        results = [TestResult.from_dict(t) for t in data.pop("results")]
+        return cls(results=results, **data, **kwargs)
 
     @property
     def failed(self) -> bool:
@@ -302,7 +291,7 @@ class CaseErrored:
     caught: bool = True
 
     def report(self, reporter: _report._CaseReporter):  # type: ignore[reportPrivateUsage]  # noqa: E501
-        reporter.errored(self)
+        reporter.case_errored(self)
 
     @classmethod
     def uncaught(
