@@ -1,10 +1,13 @@
 import { useState } from "react";
 import { Check, Clipboard } from "react-bootstrap-icons";
+import Button from "react-bootstrap/Button";
+import OverlayTrigger from "react-bootstrap/OverlayTrigger";
+import Tooltip from "react-bootstrap/Tooltip";
 
 function CopyToClipboard({ textToCopy }) {
   const [copied, setCopied] = useState(false);
 
-  const handleClick = () => {
+  const handleCopy = () => {
     navigator.clipboard.writeText(textToCopy);
     setCopied(true);
     setTimeout(() => {
@@ -13,18 +16,24 @@ function CopyToClipboard({ textToCopy }) {
   };
 
   return (
-    <button
-      type="button"
-      className="btn mt-0 me-0"
-      onClick={handleClick}
-      aria-label="Copy to clipboard"
-      data-bs-toggle="tooltip"
-      data-bs-placement="top"
-      data-bs-custom-class="custom-tooltip"
-      data-bs-title="Copy to clipboard"
-    >
-      {copied ? <Check /> : <Clipboard />}
-    </button>
+    <>
+      <OverlayTrigger
+        placement="top"
+        overlay={
+          <Tooltip id="tooltip-top">
+            {copied ? "Copied!" : "Copy to clipboard"}
+          </Tooltip>
+        }
+      >
+        <Button
+          variant="outline-primary"
+          className="btn mt-0 me-0"
+          onClick={handleCopy}
+        >
+          {copied ? <Check /> : <Clipboard />}
+        </Button>
+      </OverlayTrigger>
+    </>
   );
 }
 
